+++ shuup/admin/modules/products/views/edit_parent.py
@@ -63,7 +63,7 @@ class ProductParentBaseView(FormPartsViewMixin, UpdateView):
         if parent:
             # By default, redirect to the first parent
             return HttpResponseRedirect(
-                reverse("shuup_admin:product.edit_package", kwargs={"pk": parent.id})
+                reverse("shuup_admin:shop_product.edit_package", kwargs={"pk": parent.id})
             )
         return super(ProductParentBaseView, self).dispatch(request, *args, **kwargs)