+++ shuup/admin/modules/products/views/list.py
@@ -15,23 +15,26 @@ from shuup.admin.utils.picotable import (
     ChoicesFilter, Column, RangeFilter, TextFilter
 )
 from shuup.admin.utils.views import PicotableListView
-from shuup.core.models import Product, ProductMode
+from shuup.core.models import ProductMode, Shop, ShopProduct
 
 
 class ProductListView(PicotableListView):
-    model = Product
+    model = ShopProduct
+
     default_columns = [
         Column("primary_image", _(u"Primary Image"), display="get_primary_image",
                class_name="text-center", raw=True, ordering=1, sortable=False),
-        Column("name", _(u"Name"), sort_field="translations__name", display="name", filter_config=TextFilter(
-            filter_field="translations__name",
-            placeholder=_("Filter by name...")
-        ), ordering=2),
-        Column("sku", _(u"SKU"), display="sku", filter_config=RangeFilter(), ordering=3),
+        Column("name", _(u"Name"),
+               sort_field="product__translations__name",
+               display="product__name",
+               filter_config=TextFilter(filter_field="product__translations__name",
+                                        placeholder=_("Filter by name...")), ordering=2),
+        Column("sku", _(u"SKU"), display="product__sku", filter_config=RangeFilter(), ordering=3),
         Column("barcode", _(u"Barcode"),
-               display="barcode", filter_config=TextFilter(_("Filter by barcode...")), ordering=4),
-        Column("type", _(u"Type"), ordering=5),
-        Column("mode", _(u"Mode"), filter_config=ChoicesFilter(ProductMode.choices), ordering=6),
+               display="product__barcode", filter_config=TextFilter(_("Filter by barcode...")), ordering=4),
+        Column("type", _(u"Type"), display="product__type", ordering=5),
+        Column("mode", _(u"Mode"),
+               display="product__mode", filter_config=ChoicesFilter(ProductMode.choices), ordering=6),
     ]
 
     mass_actions = [
@@ -42,30 +45,28 @@ class ProductListView(PicotableListView):
     ]
 
     def get_primary_image(self, instance):
-        if instance.primary_image:
-            return "<img src='/media/%s'>" % instance.primary_image.get_thumbnail()
+        if instance.product.primary_image:
+            return "<img src='/media/%s'>" % instance.product.primary_image.get_thumbnail()
         else:
             return "<img src='%s'>" % static("shuup_admin/img/no_image_thumbnail.png")
 
     def get_queryset(self):
         filter = self.get_filter()
-        shop_id = filter.get("shop")
-        qs = Product.objects.all_except_deleted()
+        shop_id = filter.get("shop", Shop.objects.first().pk)
+        qs = ShopProduct.objects.filter(product__deleted=False, shop_id=shop_id)
         q = Q()
         for mode in filter.get("modes", []):
-            q |= Q(mode=mode)
+            q |= Q(product__mode=mode)
         manufacturer_ids = filter.get("manufacturers")
         if manufacturer_ids:
-            q |= Q(manufacturer_id__in=manufacturer_ids)
+            q |= Q(product__manufacturer_id__in=manufacturer_ids)
         qs = qs.filter(q)
-        if shop_id:
-            qs = qs.filter(shop_products__shop_id=int(shop_id))
         return qs
 
     def get_object_abstract(self, instance, item):
         return [
             {"text": "%s" % instance, "class": "header"},
-            {"title": _(u"Barcode"), "text": item.get("barcode")},
-            {"title": _(u"SKU"), "text": item.get("sku")},
-            {"title": _(u"Type"), "text": item.get("type")},
+            {"title": _(u"Barcode"), "text": item.get("product__barcode")},
+            {"title": _(u"SKU"), "text": item.get("product__sku")},
+            {"title": _(u"Type"), "text": item.get("product__type")},
         ]